split index.ts into individual files, no code changes
